% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{area_definition}
\alias{area_definition}
\title{ICES Area and Ecoregion definitions}
\usage{
area_definition(ecoregion = "Greater North Sea Ecoregion")
}
\arguments{
\item{ecoregion}{character of the ecoregion to plot. e.g., "Greater North Sea Ecoregion"}
}
\value{
list of ices_shape (data frame), eco_shape (data frame), europe_shape (data frame) and centroids (data frame)
}
\description{
\code{area_definition} returns list of sf dataframes with shape information of Europe, ICES Areas and ICES Ecoregions.
}
\note{
European map is from Natural Earth (scale = 10) via the rnaturalearth package
}
\examples{

}
\seealso{
Used in \code{\link{area_definition_map}} to a map describing potential mismatches between ICES Ecoregions and ICES Areas.

Input data: \code{\link{ices_shape}}, \code{\link{eco_shape}}, and \code{\link{europe_shape}}.
}
\author{
Scott Large
}
