% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_data_official}
\alias{catch_data_official}
\title{Official Nominal Catches 2006-2014}
\format{A data frame with 49182 rows and 13 variables:}
\source{
\url{http://www.ices.dk/marine-data/dataset-collections/Pages/Fish-catch-and-stock-assessment.aspx}
}
\usage{
catch_data_official
}
\description{
Catches in FAO area 27 by country, species, area and year.
Source: Eurostat/ICES database on catch statistics - ICES 2011, Copenhagen.
Version 30-11-2011

\itemize{
\item{Species}{Add text}
\item{Area}{Add text}
\item{Units}{Add text}
\item{Country}{Add text}
\item{X2014}{Add text}
\item{X2013}{Add text}
\item{X2012}{Add text}
\item{X2011}{Add text}
\item{X2010}{Add text}
\item{X2009}{Add text}
\item{X2008}{Add text}
\item{X2007}{Add text}
\item{X2006}{Add text}
}
}
\keyword{datasets}
