% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{clean_sag}
\alias{clean_sag}
\title{Clean SAG reference points and summary table}
\usage{
clean_sag(active_year = 2016)
}
\arguments{
\item{active_year}{numeric of the stock database version. e.g., 2016}
}
\value{
list of stock_list_frmt (data frame) and sag_complete_summary (tbl_df)
}
\description{
\code{clean_sag} returns a merged and tidied SAG reference points and summary table and a formatted (.html) stock list
}
\note{
Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.
}
\examples{
head(clean_sag(active_year = 2016))
}
\seealso{
Used in \code{\link{stock_trends_fun}} to make clickable information on dynamic
line plots of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion.
Also, used in \code{\link{frmt_summary_table}}.

Input data: \code{\link{stock_list_raw}}, \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys_raw}}.
}
\author{
Scott Large
}
