% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{clean_stock_trends}
\alias{clean_stock_trends}
\title{Clean stock status trends}
\usage{
clean_stock_trends(active_year = 2016, grouping_var = c("EcoGuild",
  "EcoRegion", "FisheriesGuild")[1], plotting_var = c("StockCode",
  "FisheriesGuild")[1], metric = c("MSY", "MEAN")[1])
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{grouping_var}{character string of the desired grouping. Options include: EcoRegion, EcoGuild, or FisheriesGuild}

\item{plotting_var}{character string of the variable to plot. Options include: StockCode or FisheriesGuild (mean)}

\item{metric}{character string of the desired metric. Options include: MSY or MEAN (according to grouping_var)}
}
\value{
A list of 2: 1) data frame of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub> reference points
for stocks of a fish category for an ecoregion and 2) \code{sag_complete_summary} from \code{\link{clean_sag}}.
}
\description{
\code{clean_stock_trends} returns list of 2: data frame of F and SSB relative to
 F<sub>MSY</sub> and MSY B<sub>trigger</sub> reference points for stocks of a fish category for an
 ecoregion and \code{sag_complete_summary} from \code{\link{clean_sag}}.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
\dontrun{
clean_stock_trends(2016)
}
}
\seealso{
Used in \code{\link{stock_trends_fun}} for line plots of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion.
Input data: From \code{\link{clean_sag}}
}
\author{
Scott Large
}
