% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{col_oxygen}
\alias{col_oxygen}
\title{Internal function to generate roxygen formatting for data files.}
\usage{
col_oxygen(object)
}
\arguments{
\item{base_num}{The number to multiply by three}
}
\value{
Returns a string ready to copy and paste into Roxygen
}
\description{
Internal function to generate roxygen formatting for data files.
}
\keyword{internal}
