% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{eco_shape}
\alias{eco_shape}
\title{ICES Ecoregions}
\format{A sf data frame with 17 rows and 6 variables.}
\source{
\url{http://gis.ices.dk/shapefiles/ICES_ecoregions.zip}
}
\usage{
eco_shape
}
\description{
ICES Ecoregions. Accessed 23 May 2017.
\itemize{
\item{OBJECTID}{add text}
\item{Ecoregion}{add text}
\item{Shape_Leng}{add text}
\item{Shape_Le_1}{add text}
\item{Shape_Area}{add text}
\item{geometry}{add text}
}
}
\keyword{datasets}
