% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{europe_shape}
\alias{europe_shape}
\title{Europe map}
\format{A sf data frame with 51 rows and 6 variables. From Natural Earth via the rnaturalearth package}
\source{
\url{http://www.naturalearthdata.com/}, \url{https://cran.r-project.org/package=rnaturalearth}
}
\usage{
europe_shape
}
\description{
ICES Ecoregions. Accessed 23 May 2017.
\itemize{
\item{iso_a3}{add text}
\item{iso_n3}{add text}
\item{admin}{add text}
\item{geometry}{add text}
}
}
\keyword{datasets}
