% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{ges_stock_props}
\alias{ges_stock_props}
\title{Proportion of stocks relative to GES reference points}
\usage{
ges_stock_props(active_year = active_year, ecoregion = ecoregion,
  calculate_status = calculate_status, fisheries_guild = fisheries_guild)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{ecoregion}{vector of ecoregions to include}

\item{calculate_status}{logical on whether to use raw SAG output to calculate stock status or to use the hard-coded values from stock summary table}

\item{fisheries_guild}{vector of fisheries guilds to include}
}
\value{
returns a data frame of the proportion of stocks relative
to ICES reference points for fish categories in ecoregions.
}
\description{
\code{ges_stock_props} returns a data frame of the proportion of stocks
 relative to GES reference points for all ecoregion.
}
\note{
Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.\code{calculate_status = TRUE} calculates stock status
relative to published reference points. This will represent PA and SBL for ecoregions with proxy reference points. \code{calculate_status = TRUE} takes the
raw icons from published advice. Note, before 2017 not all stocks status tables have been added to the SAG database and only few stocks had MSY proxy reference points.
}
\examples{
\dontrun{
ges_stock_props(2016)
}
}
\seealso{
Used in \code{\link{gesPie_fun}} to plot proportion of stocks relative
to GES reference points for all ecoregions.
Input data: From \code{\link{stock_props}}, which brings \code{sag_complete_summary} and \code{stock_list_frmt}
from \code{\link{clean_sag}}.
}
\author{
Scott Large
}
