% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{ges_table}
\alias{ges_table}
\title{ges_table}
\format{A data frame with 128 rows and 5 variables.}
\usage{
ges_table
}
\description{
Proportion of stocks and landings relative to GES reference points in an ecoregion.
}
\details{
\itemize{
\item{EcoRegion}{Add text}
\item{VARIABLE}{Add text}
\item{COLOR}{Add text}
\item{VALUE}{Add text}
\item{METRIC}{Add text}
}
}
\seealso{
Used in \code{\link{gesPie_fun}} to plot dynamic and
 static tables of stock status by ecoregion. Input data: Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
