% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{gesPie_fun}
\alias{gesPie_fun}
\title{Pie chart of proportion of stocks relative to GES reference points}
\usage{
gesPie_fun(ecoregion, fisheries_guild = c("pelagic", "demersal", "crustacean",
  "elasmobranch", "benthic"), calculate_status = FALSE, data_caption = TRUE,
  file_name = NULL, active_year = 2016, save_plot = FALSE,
  return_plot = TRUE, return_data = FALSE, output_path = NULL)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea}

\item{fisheries_guild}{fisheries guild to include in proportions}

\item{calculate_status}{logical on whether to use raw SAG output to calculate stock status or to use the hard-coded values from stock summary table}

\item{data_caption}{print the data source as a caption, boolean.}

\item{file_name}{name for the output.}

\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{save_plot}{logical to save plot.}

\item{return_plot}{logical to return plot to current environment.}

\item{return_data}{logical on returning a .csv of plotted data}

\item{output_path}{path for output to live.}
}
\value{
A ggplot2 object or .png saved as \code{file_name} to \code{output_path}.
When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/".
}
\description{
The \code{gesPie_fun} function returns pie charts of the proportion of stocks
relative to GES reference points in an ecoregion.
}
\note{
Stocks are linked to ecoregions via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
 when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20). \code{calculate_status = TRUE} calculates stock status
relative to published reference points. This will represent PA and SBL for ecoregions with proxy reference points. \code{calculate_status = TRUE} takes the
raw icons from published advice. Note, before 2017 not all stocks status tables have been added to the SAG database and only few stocks had MSY proxy reference points.
}
\examples{
\dontrun{
gesPie_fun("Greater North Sea Ecoregion", return_plot = TRUE)
}
}
\author{
Scott Large
}
