% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{guild_discards_fun}
\alias{guild_discards_fun}
\title{Discard rate over time}
\usage{
guild_discards_fun(ecoregion, active_year = 2016, data_caption = TRUE,
  output_path = NULL, save_plot = FALSE, return_plot = TRUE,
  file_name = NULL)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea Ecoregion}

\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{data_caption}{print the data source as a caption, boolean.}

\item{output_path}{path for output to live.}

\item{save_plot}{logical to save plot.}

\item{return_plot}{logical to return plot to current environment.}

\item{file_name}{name for the output.}
}
\value{
A ggplot2 object when \code{return_plot} is \code{TRUE}, html when \code{dynamic} is \code{TRUE}
or .png when \code{dynamic} is \code{FALSE}. Output is saved as \code{file_name} in \code{output_path}.
When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/".
}
\description{
This function returns a series of plots of discard rate and landings by fish category for an ecoregion.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
There is an assumption that discard rates for biannual stocks and
 are consistent over the years that we don't provide new advice.
}
\examples{
\dontrun{
guild_discards_fun("Greater North Sea Ecoregion", return_plot = TRUE)
}
}
\author{
Scott Large
}
