% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{ices_catch_data}
\alias{ices_catch_data}
\title{ICES catch time series}
\usage{
ices_catch_data()
}
\value{
a data frame of landings by country, species, and fish category.
}
\description{
\code{ices_catch_data} returns a data frame of landings by country, species, and fish category
}
\examples{
head(ices_catch_data())
}
\seealso{
Used in \code{\link{ices_catch_plot}} to create an area or line plot of landings (historic and official catch)
for an ecoregion by country, fish category, or species.
Input data: \code{\link{catch_data_official}}, \code{\link{catch_data_historical}}, \code{\link{species_list}},
and \code{\link{stock_list_raw}}.
}
\author{
Scott Large
}
