% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{ices_catch_historical_raw}
\alias{ices_catch_historical_raw}
\title{Historical Nominal Catches 1950-2010}
\format{A data frame with 28582 rows and 64 variables:}
\source{
\url{http://www.ices.dk/marine-data/dataset-collections/Pages/Fish-catch-and-stock-assessment.aspx}
}
\usage{
ices_catch_historical_raw
}
\description{
Catches in FAO area 27 by country, species, area and year as provided
by the national authorities.
Source: Eurostat/ICES data compilation of catch statistics - ICES 2011, Copenhagen.
Version: 30-11-2011
}
\details{
\itemize{
\item{Country}{Add text}
\item{Species}{Add text}
\item{Division}{Add text}
\item{X1950}{Add text}
 ...
\item{X2010}{Add text}
}
}
\keyword{datasets}
