% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{ices_catch_official_raw}
\alias{ices_catch_official_raw}
\title{Official Nominal Catches 2006-2015}
\format{A data frame with 50203 rows and 14 variables:}
\source{
\url{http://www.ices.dk/marine-data/dataset-collections/Pages/Fish-catch-and-stock-assessment.aspx}
}
\usage{
ices_catch_official_raw
}
\description{
Catches in FAO area 27 by country, species, area and year.
Source: Eurostat/ICES database on catch statistics - ICES 2017, Copenhagen.
Version 12-06-2017

\itemize{
\item{Species}{Add text}
\item{Area}{Add text}
\item{Units}{Add text}
\item{Country}{Add text}
\item{X2015}{Add text}
\item{X2014}{Add text}
\item{X2013}{Add text}
\item{X2012}{Add text}
\item{X2011}{Add text}
\item{X2010}{Add text}
\item{X2009}{Add text}
\item{X2008}{Add text}
\item{X2007}{Add text}
\item{X2006}{Add text}
}
}
\keyword{datasets}
