% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ices_catch_plot}
\alias{ices_catch_plot}
\title{Landings over time by country, guild, or species}
\usage{
ices_catch_plot(ecoregion, type = c("COMMON_NAME", "COUNTRY", "GUILD")[1],
  line_count = 4, plot_type = c("line", "area")[1], data_caption = TRUE,
  output_path = NULL, file_name = "figure2", save_plot = FALSE,
  return_plot = TRUE, fig.width = 174, fig.height = 68, text.size = 9)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea Ecoregion}

\item{type}{the variable that will be used to group and display data: COMMON_NAME, GUILD, or COUNTRY}

\item{line_count}{number of lines to display}

\item{plot_type}{area or line plot}

\item{data_caption}{print the data source as a caption, boolean.}

\item{output_path}{path for output to live.}

\item{file_name}{name for the output.}

\item{save_plot}{logical to save plot.}

\item{return_plot}{logical to return plot to current environment.}

\item{fig.width}{width pf combined set of plots}

\item{fig.height}{height of combined set of plots}

\item{text.size}{= size of text in plots}
}
\value{
A ggplot2 object when \code{return_plot} is \code{TRUE} or .png when \code{save_plot} is \code{TRUE}.
Output is saved as \code{file_name} in \code{output_path}.
When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/".
}
\description{
This function returns an area or line plot of landings (historic and official catch) for an ecoregion by country,
guild, or species.
}
\note{
Historic and official nominal catch are actually only the landings and do not account for discards, misreporting, or other
potential issues.
}
\examples{
\dontrun{
ices_catch_plot("Greater North Sea Ecoregion", type = "COMMON_NAME", return_plot = TRUE, line_count = 4)
}
}
\author{
Scott Large
}
