% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{ices_shape}
\alias{ices_shape}
\title{ICES Statistical Areas}
\format{A sf data frame with 66 rows and 11 variables.}
\source{
\url{http://gis.ices.dk/shapefiles/ICES_areas.zip}
}
\usage{
ices_shape
}
\keyword{datasets}
