% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{pa_maker}
\alias{pa_maker}
\title{Convert FPA to SBL}
\usage{
pa_maker(df, assessment_year)
}
\arguments{
\item{df}{numeric of the stock database version. e.g., 2016}

\item{assessment_year}{logical to return objects from clean_sag()}
}
\value{
data frame
}
\description{
\code{pa_maker} helper function that takes FPA and BPA and converts to SB.
}
\note{
internal function
}
\seealso{
Used in \code{\link{frmt_summary_tbl}} to create the "Status of stock summary relative to reference points"
table for all stocks for an ecoregion.
}
\author{
Scott Large
}
\keyword{internal}
