% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{pie_table_count}
\alias{pie_table_count}
\title{pie_table_count}
\format{A data frame with 1189 rows and 5 variables.}
\usage{
pie_table_count
}
\description{
Proportion of stocks relative to reference points for fish categories in an ecoregion.
}
\details{
\itemize{
\item{EcoRegion}{Ecoregion}
\item{FisheriesGuild}{Fisheries category. e.g., benthic, demersal, pelagic, crustacean, elasmobranch, or undetermined}
\item{VARIABLE}{Reference point}
\item{VALUE}{Categorical value of status relative to reference points. e.g., GREEN, RED, GREY, or ORANGE}
\item{COUNT}{Number of stocks that are within each ecoregion, fisheries category, and variable combination}
}
}
\seealso{
Used in \code{\link{stockPie_fun}} to plot proportion of stocks relative
 to reference points for fish categories in an ecoregion. Input data: Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
