% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{sag_complete_summary}
\alias{sag_complete_summary}
\title{sag_complete_summary}
\format{A tbl_df with 7977 rows and 25 variables.}
\usage{
sag_complete_summary
}
\description{
Merged and tidied SAG reference points and summary table. Used to make clickable information on dynamic plots
}
\details{
\itemize{
\item{StockCode}{Add text}
\item{Description}{Add text}
\item{SpeciesScientificName}{Add text}
\item{DataCategory}{Add text}
\item{YearOfLastAssessment}{Add text}
\item{AdviceCategory}{Add text}
\item{FisheriesGuild}{Add text}
\item{SpeciesID}{Add text}
\item{Flim}{Add text}
\item{Fpa}{Add text}
\item{Bpa}{Add text}
\item{Blim}{Add text}
\item{FMSY}{Add text}
\item{MSYBtrigger}{Add text}
\item{Year}{Add text}
\item{F}{Add text}
\item{SSB}{Add text}
\item{fishingPressureDescription}{Add text}
\item{stockSizeDescription}{Add text}
\item{landings}{Add text}
\item{catches}{Add text}
\item{discards}{Add text}
\item{FmsyDescription}{Add text}
\item{BmsyDescription}{Add text}
\item{data}{Add text}
}
}
\seealso{
Used in \code{\link{stock_trends_fun}}  to make clickable information on dynamic
line plots of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion. Input data: Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
