% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{sag_keys_raw}
\alias{sag_keys_raw}
\title{ICES Stock Assessment Graphs database - keys}
\format{A data frame with 545 rows and 3 variables:}
\source{
\url{https://standardgraphs.ices.dk/}
}
\usage{
sag_keys_raw
}
\description{
Data from published ICES advice from 2014-2017. Accessed Accessed 5 July 2017. “ICES Stock Assessment Database, 2017/July ICES, Copenhagen”
}
\details{
\itemize{
\item{AssessmentYear}{Add text}
\item{AssessmentKey}{Add text}
\item{StockKeyLabel}{Add text}
}
}
\keyword{datasets}
