% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{sag_refpts_raw}
\alias{sag_refpts_raw}
\title{ICES Stock Assessment Graphs database - reference points}
\format{A data frame with 545 rows and 15 variables:}
\source{
\url{https://standardgraphs.ices.dk/}
}
\usage{
sag_refpts_raw
}
\description{
Data from published ICES advice from 2014-2017. Accessed Accessed 5 July 2017. “ICES Stock Assessment Database, 2017/July ICES, Copenhagen”
}
\details{
\itemize{
\item{AssessmentKey}{Add text}
\item{StockKeyLabel}{Add text}
\item{StockDatabaseID}{Add text}
\item{StockKey}{Add text}
\item{AssessmentYear}{Add text}
\item{FLim}{Add text}
\item{Fpa}{Add text}
\item{Bpa}{Add text}
\item{Blim}{Add text}
\item{FMSY}{Add text}
\item{MSYBtrigger}{Add text}
\item{Fmanagement}{Add text}
\item{Bmanagement}{Add text}
\item{RecruitmentAge}{Add text}
\item{RecruitmentLength}{Add text}
}
}
\keyword{datasets}
