% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{sag_stock_status_raw}
\alias{sag_stock_status_raw}
\title{ICES Stock Assessment Graphs database - stock status output}
\format{A data frame with 6163 rows and 13 variables:}
\source{
\url{https://standardgraphs.ices.dk/}
}
\usage{
sag_stock_status_raw
}
\description{
Data from published ICES advice from 2014-2017. Accessed Accessed 5 July 2017. “ICES Stock Assessment Database, 2017/July. ICES, Copenhagen”
}
\details{
\itemize{
\item{AssessmentYear}{Add text}
\item{AssessmentKey}{Add text}
\item{StockKeyLabel}{Add text}
\item{year}{Add text}
\item{status}{Add text}
\item{statusicon}{Add text}
\item{type}{Add text}
\item{lineNumber}{Add text}
\item{lineDescription}{Add text}
\item{fishingPressure}{Add text}
\item{stockSize}{Add text}
\item{stockSizeStatus}{Add text}
\item{fishingPressureStatus}{Add text}
}
}
\keyword{datasets}
