% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{sag_summary_raw}
\alias{sag_summary_raw}
\title{ICES Stock Assessment Graphs database - summary information from assessment output}
\format{A data frame with 20204 rows and 23 variables:}
\source{
\url{https://standardgraphs.ices.dk/}
}
\usage{
sag_summary_raw
}
\description{
Data from published ICES advice from 2014-2017. Accessed Accessed 5 July 2017. ICES Stock Assessment Database, 2017/July ICES, Copenhagen
}
\details{
\itemize{
\item{Year}{Add text}
\item{recruitment}{Add text}
\item{high_recruitment}{Add text}
\item{low_recruitment}{Add text}
\item{low_SSB}{Add text}
\item{SSB}{Add text}
\item{high_SSB}{Add text}
\item{catches}{Add text}
\item{landings}{Add text}
\item{discards}{Add text}
\item{low_F}{Add text}
\item{F}{Add text}
\item{high_F}{Add text}
\item{StockPublishNote}{Add text}
\item{Fage}{Add text}
\item{fishstock}{Add text}
\item{recruitment_age}{Add text}
\item{AssessmentYear}{Add text}
\item{units}{Add text}
\item{stockSizeDescription}{Add text}
\item{stockSizeUnits}{Add text}
\item{fishingPressureDescription}{Add text}
\item{fishingPressureUnits}{Add text}
}
}
\keyword{datasets}
