% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{species_list_raw}
\alias{species_list_raw}
\title{ASFIS list of species}
\format{A data frame with 12700 rows and 11 variables:}
\source{
\url{http://www.fao.org/fishery/collection/asfis/en}
}
\usage{
species_list_raw
}
\description{
ASFIS list of species includes 12 700 species items selected according
to their interest or relation to fisheries and aquaculture.
For each species item stored in a record, codes (ISSCAAP group, taxonomic and 3-alpha)
and taxonomic information (scientific name, author(s), family, and higher taxonomic classification) are provided.
}
\details{
Version 2-2016

\itemize{
\item{ISSCAAP}{Add text}
\item{TAXOCODE}{Add text}
\item{X3A_CODE}{Add text}
\item{Scientific_name}{Add text}
\item{English_name}{Add text}
\item{French_name}{Add text}
\item{Spanish_name}{Add text}
\item{Author}{Add text}
\item{Family}{Add text}
\item{Order}{Add text}
\item{Stats_data}{Add text}
}
}
\keyword{datasets}
