% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{stecf_data}
\alias{stecf_data}
\title{STECF catch and effort time series}
\usage{
stecf_data()
}
\value{
a data frame of stock status relative to reference points and catch, discards, and landings
by stock for the most recent assessment.
}
\description{
\code{stecf_data} returns a data frame of effort and landings by country, species, and fish category
}
\examples{
head(stecf_data())
}
\seealso{
Used in \code{\link{stecf_plot}} to create an area or line plot of landings and effort for an ecoregion
by country and guild.
Input data: \code{\link{stecf_effort_raw}} and \code{\link{stecf_landings_raw}}.
}
\author{
Scott Large
}
