% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STECF_effort_data}
\alias{STECF_effort_data}
\title{STECF nominal effort}
\format{A data frame with 23993 rows and 13 variables:}
\source{
\url{https://stecf.jrc.ec.europa.eu/dd/effort/graphs-annex}
}
\usage{
STECF_effort_data
}
\description{
STECF nominal effort from the FDI data call 2016. Accessed 16 March 2017.
}
\keyword{datasets}
