% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{stecf_effort_raw}
\alias{stecf_effort_raw}
\title{STECF nominal effort}
\format{A data frame with 23993 rows and 13 variables.}
\source{
\url{https://stecf.jrc.ec.europa.eu/dd/effort/graphs-annex}
}
\usage{
stecf_effort_raw
}
\description{
STECF nominal effort from the FDI data call 2016. Accessed 16 March 2017.
}
\details{
\itemize{
\item{measure.calculation}{add text}
\item{annex}{add text}
\item{country}{add text}
\item{fishing_activity}{add text}
\item{fishing_capacity}{add text}
\item{gt_days_at_sea}{add text}
\item{no_vessels}{add text}
\item{nominal_effort}{add text}
\item{regulated.area}{add text}
\item{regulated.gear}{add text}
\item{specon}{add text}
\item{vessel.length}{add text}
\item{year}{add text}
}
}
\keyword{datasets}
