% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{stecf_landings_clean}
\alias{stecf_landings_clean}
\title{stecf_landings_clean}
\format{A data frame with 443077 rows and 9 variables.}
\usage{
stecf_landings_clean
}
\description{
Historic and official catch for an ecoregion by country, guild, and species.
}
\details{
\itemize{
\item{YEAR}{Add text}
\item{COUNTRY}{Add text}
\item{ISO3}{Add text}
\item{GUILD}{Add text}
\item{ECOREGION}{Add text}
\item{SPECIES_NAME}{Add text}
\item{SPECIES_CODE}{Add text}
\item{COMMON_NAME}{Add text}
\item{VALUE}{Add text}
}
}
\seealso{
Used in \code{\link{ices_catch_plot}} to create an area or line plot of landings (historic and official catch) for an ecoregion by country,
guild, or species.
Input data: Input data: \code{\link{catch_data_official}}, \code{\link{catch_data_historical}}, \code{\link{species_list}}, and \code{\link{stock_list_raw}}.
}
\keyword{datasets}
