% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STECF_landings_data}
\alias{STECF_landings_data}
\title{STECF landings and discards}
\format{A data frame with 530230 rows and 11 variables:}
\source{
\url{https://stecf.jrc.ec.europa.eu/dd/effort/graphs-annex}
}
\usage{
STECF_landings_data
}
\keyword{datasets}
