% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{stecf_landings_raw}
\alias{stecf_landings_raw}
\title{STECF landings and discards}
\format{A data frame with 530230 rows and 11 variables.}
\source{
\url{https://stecf.jrc.ec.europa.eu/dd/effort/graphs-annex}
}
\usage{
stecf_landings_raw
}
\keyword{datasets}
