% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{stock_catch}
\alias{stock_catch}
\title{Catch, discards, and landings by stock}
\usage{
stock_catch(active_year = 2016)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}
}
\value{
a data frame of reference points, catch, discards, and landings by stock over time.
}
\description{
\code{stock_catch} returns a data frame of reference points, catch, discards, and landings by stock over time.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
head(stock_catch(2016))
}
\seealso{
Used in \code{\link{guild_discards_fun}} for plots of discard rate and landings by fish category for an ecoregion.
Input data: From \code{\link{clean_sag}}
}
\author{
Scott Large
}
