% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{stock_list_raw}
\alias{stock_list_raw}
\title{ICES Stock database}
\format{A data frame with 1241 rows and 31 variables.}
\source{
\url{https://sd.ices.dk/}
}
\usage{
stock_list_raw
}
\description{
Data describing ICES Stocks. Accessed 13 June 2017
}
\details{
\itemize{
 \item{StockDatabaseID}{Add text}
 \item{StockKey}{Add text}
 \item{StockKeyLabel}{Add text}
 \item{StockKeyDescription}{Add text}
 \item{PreviousStockKey}{Add text}
 \item{PreviousStockKeyLabel}{Add text}
 \item{ActiveYear}{Add text}
 \item{SpeciesScientificName}{Add text}
 \item{SpeciesCommonName}{Add text}
 \item{EcoRegion}{Add text}
 \item{ExpertGroup}{Add text}
 \item{ExpertGroupDescription}{Add text}
 \item{AdviceDraftingGroup}{Add text}
 \item{AdviceDraftingGroupDescription}{Add text}
 \item{DataCategory}{Add text}
 \item{YearOfLastAssessment}{Add text}
 \item{AssessmentFrequency}{Add text}
 \item{YearOfNextAssessment}{Add text}
 \item{AssessmentType}{Add text}
 \item{AdviceReleaseDate}{Add text}
 \item{AdviceCategory}{Add text}
 \item{AdviceType}{Add text}
 \item{UseOfDiscardsInAdvice}{Add text}
 \item{PABufferApplied}{Add text}
 \item{TrophicGuild}{Add text}
 \item{FisheriesGuild}{Add text}
 \item{SizeGuild}{Add text}
 \item{Published}{Add text}
 \item{GeneratedOn}{Add text}
 \item{SectionNumber}{Add text}
 \item{AssessmentKey}{Add text}
}
}
\keyword{datasets}
