% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{stock_props}
\alias{stock_props}
\title{Proportion of stocks relative to reference points}
\usage{
stock_props(active_year, ecoregion, fisheries_guild, return_clean_sag = FALSE)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{ecoregion}{vector of ecoregions to include}

\item{fisheries_guild}{vector of fisheries guilds to include}

\item{return_clean_sag}{logical to return objects from clean_sag()}
}
\value{
returns a list with the data frame \code{stock_props} of the proportion of stocks relative
to reference points for all ecoregions. When \code{return_clean_sag = TRUE}, \code{sag_complete_summary} and \code{stock_list_frmt}
are also returned from \code{\link{clean_sag}}.
}
\description{
\code{stock_props} returns a list of the proportion of stocks
 relative to reference points for fish categories in ecoregions.
}
\note{
Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.
}
\examples{
\dontrun{
stock_props(2016)
}
}
\seealso{
Used in \code{\link{ices_stock_props}} and \code{\link{ges_stock_props}} to evaluate stocks relative
to ICES and GES reference points.
Input data: SAG summary table and reference points come from \code{\link{clean_sag}}. \code{\link{frmt_summary_table}} evaluates
status relative to reference points and formats the table for .html.
}
\author{
Scott Large
}
