% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{stock_status}
\alias{stock_status}
\title{Stock status and catch relative to reference points}
\usage{
stock_status(active_year = 2016)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}
}
\value{
a data frame of stock status relative to reference points and catch, discards, and landings
by stock for the most recent assessment.
}
\description{
\code{stock_status} returns a data frame of stock status relative to reference points and
catch, discards, and landings by stock for the most recent assessment.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
head(stock_status(2016))
}
\seealso{
Used in \code{\link{plot_kobe}} to plot a scatter plot of F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion and a "lollipop" plot of total catch (divided into discards and landings) by stock.
Input data: From \code{\link{stock_catch}}
}
\author{
Scott Large
}
