% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{stock_status_full}
\alias{stock_status_full}
\title{stock_status_full}
\format{A data frame with 759 rows and 14 variables.}
\usage{
stock_status_full
}
\description{
F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub> by fish category and ecoregion and
total catch (divided into discards and landings) by stock.
}
\details{
\itemize{
\item{StockCode}{Add text}
\item{Description}{Add text}
\item{FisheriesGuild}{Add text}
\item{EcoRegion}{Add text}
\item{F_FMSY}{Add text}
\item{catches}{Add text}
\item{landings}{Add text}
\item{discards}{Add text}
\item{FMSY}{Add text}
\item{F}{Add text}
\item{SSB_MSYBtrigger}{Add text}
\item{SSB}{Add text}
\item{MSYBtrigger}{Add text}
\item{colList}{Add text}
}
}
\seealso{
Used in \code{\link{plot_kobe}} to 2 plots: a scatter plot of F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion and a "lollipop" plot of total catch (divided into discards and landings) by stock.
Input data: Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
