% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{stock_trends_frmt}
\alias{stock_trends_frmt}
\title{stock_trends_frmt}
\format{A data frame with 19692 rows and 5 variables.}
\usage{
stock_trends_frmt
}
\description{
F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion.
}
\details{
\itemize{
\item{pageGroup}{Add text}
\item{lineGroup}{Add text}
\item{Year}{Add text}
\item{plotGroup}{Add text}
\item{plotValue}{Add text}
}
}
\seealso{
Used in \code{\link{stock_trends_fun}} to plot line plots of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion. Input data: Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
