% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{stock_trends_fun}
\alias{stock_trends_fun}
\title{Stock status over time}
\usage{
stock_trends_fun(EcoGuild, active_year = 2016, dynamic = TRUE,
  data_caption = TRUE, file_name = NULL, save_plot = FALSE,
  return_plot = TRUE, return_data = FALSE, output_path = NULL,
  stackable = FALSE)
}
\arguments{
\item{EcoGuild}{combined ecoregion name and fish category, e.g. Greater North Sea Ecoregion}

\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{dynamic}{logical to generate html output with dynamic features.}

\item{data_caption}{print the data source as a caption, boolean.}

\item{file_name}{name for the output.}

\item{save_plot}{logical to save plot.}

\item{return_plot}{logical to return plot to current environment.}

\item{return_data}{logical on returning a .csv of plotted data}

\item{output_path}{path for output to live.}
}
\value{
A ggplot2 object when \code{return_plot} is \code{TRUE}, html when \code{dynamic} is \code{TRUE}
or .png when \code{dynamic} is \code{FALSE}. Output is saved as \code{file_name} in \code{output_path}.
When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/".
}
\description{
The \code{stock_trends_fun} function returns a series of line plots of F and SSB relative to F<sub>MSY</sub> and MSY B<sub>trigger</sub>
reference points for stocks of a fish category for an ecoregion.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
\dontrun{
stock_trends_fun("Greater North Sea Ecoregion - demersal", return_plot = TRUE)
}
}
\author{
Scott Large
}
