% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{stockPie_fun}
\alias{stockPie_fun}
\title{Pie chart of proportion of stocks relative to reference points}
\usage{
stockPie_fun(ecoregion, file_name, save_plot = FALSE, return_plot = TRUE,
  output_path = NULL)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea}

\item{file_name}{name for the output.}

\item{save_plot}{logical to save plot.}

\item{return_plot}{logical to return plot to current environment.}

\item{output_path}{path for output to live.}
}
\value{
A ggplot2 object or .png saved as \code{file_name} to \code{output_path}.
When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/".
}
\description{
This function returns pie charts of the proportion of stocks
relative to reference points for fish categories in an ecoregion.
}
\note{
Stocks are linked to ecoregions and fish categories via the ICES Stock database.
Reference points are as published in ICES Stock Assessment Graphs database. In some cases,
status relative to reference points may vary from published ICES advice
 when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
\dontrun{
stockPie_fun("Greater North Sea Ecoregion", return_plot = TRUE)
}
}
\author{
Scott Large
}
