% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{stockSummaryTable_fun}
\alias{stockSummaryTable_fun}
\title{Render html stock summary table}
\usage{
stockSummaryTable_fun(ecoregion, active_year = 2016,
  table_type = c("static_docx", "dynamic_html")[1], output_path = NULL,
  file_name = NULL)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea}

\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{table_type}{type of table, "dynamic_html" (using DT) and "static_docx" (using \code{flextable} and \code{officer}) .docx tables.}

\item{output_path}{path for output to live.}

\item{file_name}{name for the output.}
}
\value{
A html file. When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/". When \code{table_type} is \code{"static"} or \code{"both"},
it might take a bit of time...
}
\description{
This function returns "Status of stock summary relative to reference points" for all stocks
in an ecoregion.
}
\note{
Stocks are linked to ecoregions via the ICES Stock database. Reference points are as published in ICES Stock Assessment
Graphs database. In some cases, status relative to reference points may vary from
published ICES advice when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).

Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.
}
\examples{
\dontrun{
stockSummaryTable_fun("Greater North Sea Ecoregion")
}
}
\seealso{
SAG summary table and reference points come from \code{\link{clean_sag}}. \code{\link{frmt_summary_table}} evaluates
status relative to reference points and formats the table for .html.
}
\author{
Scott Large
}
