% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate_data.R
\docType{data}
\name{summary_table_frmt}
\alias{summary_table_frmt}
\title{summary_table_frmt}
\format{A data frame with 799 rows and 16 variables.}
\usage{
summary_table_frmt
}
\description{
Summary table
}
\details{
\itemize{
\item{StockCode}{Add text}
\item{Description}{Add text}
\item{FisheriesGuild}{Add text}
\item{EcoRegion}{Add text}
\item{AdviceCategory}{Add text}
\item{DataCategory}{Add text}
\item{SBL}{Add text}
\item{F_2013}{Add text}
\item{F_2014}{Add text}
\item{F_2015}{Add text}
\item{SSB_2014}{Add text}
\item{SSB_2015}{Add text}
\item{SSB_2016}{Add text}
\item{D3C1}{Add text}
\item{D3C2}{Add text}
\item{GES}{Add text}
}
}
\seealso{
Used in \code{\link{stockPie_fun}} to plot pie charts of the proportion
of st stock status by ecoregion. Input data: \code{\link{stock_list_raw}},
 \code{\link{sag_summary_raw}},\code{\link{sag_refpts_raw}}, and \code{\link{sag_keys}}.
}
\keyword{datasets}
