% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{ges_stock_props}
\alias{ges_stock_props}
\title{Proportion of stocks relative to GES reference points}
\usage{
ges_stock_props(active_year, ecoregion, fisheries_guild)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}

\item{ecoregion}{vector of ecoregions to include}

\item{fisheries_guild}{vector of fisheries guilds to include}
}
\value{
returns a data frame of the proportion of stocks relative
to ICES reference points for fish categories in ecoregions.
}
\description{
\code{ges_stock_props} returns a data frame of the proportion of stocks
 relative to GES reference points for all ecoregion.
}
\note{
Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.
}
\examples{
\dontrun{
ges_stock_props(2016)
}
}
\seealso{
Used in \code{\link{gesPie_fun}} to plot proportion of stocks relative
to GES reference points for all ecoregions.
Input data: From \code{\link{stock_props}}, which brings \code{sag_complete_summary} and \code{stock_list_frmt}
from \code{\link{clean_sag}}.
}
\author{
Scott Large
}
