% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{ices_stock_props}
\alias{ices_stock_props}
\title{Proportion of stocks relative to ICES reference points}
\usage{
ices_stock_props(active_year)
}
\arguments{
\item{active_year}{numeric of the stock database version (year). e.g., 2016}
}
\value{
returns a data frame of the proportion of stocks relative
to ICES reference points for fish categories in ecoregions.
}
\description{
\code{ices_stock_props} returns a data frame of the proportion of stocks
 relative to ICES reference points for fish categories in ecoregions.
}
\note{
Periodically, ICES adds or removes stocks from the advisory process. The function returns
the SAG reference points and summary table for all published (in SAG) and active stocks for a given year.
}
\examples{
\dontrun{
ices_stock_props(2016)
}
}
\seealso{
Used in \code{\link{stockPie_fun}} to plot proportion of stocks relative
to ICES reference points for fish categories in all ecoregions.
Input data: From \code{\link{stock_props}}.
}
\author{
Scott Large
}
