% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003-plot.R
\name{stockSummaryTable_fun}
\alias{stockSummaryTable_fun}
\title{Render html stock summary table}
\usage{
stockSummaryTable_fun(ecoregion, table_type = c("static", "dynamic",
  "both")[2], output_path = NULL, file_name = NULL)
}
\arguments{
\item{ecoregion}{ecoregion name, e.g. Greater North Sea}

\item{table_type}{type of table, "both" returns both "dynamic" (using DT) and "static" (using ReporteRs) html tables.}

\item{output_path}{path for output to live.}

\item{file_name}{name for the output.}
}
\value{
A html file. When \code{file_name} is \code{NULL}, the file name is the ecoregion.
When \code{output_path} is \code{NULL}, the file is saved to "~/". When \code{table_type} is \code{"static"} or \code{"both"},
it might take a bit of time...
}
\description{
This function returns "Status of stock summary relative to reference points" for all stocks
in an ecoregion.
}
\note{
Stocks are linked to ecoregions via the ICES Stock database. Reference points are as published in ICES Stock Assessment
Graphs database. In some cases, status relative to reference points may vary from
published ICES advice when reported F or SSB are very close to reference points (e.g., F = 0.201 > F<sub>MSY</sub> = 0.20).
}
\examples{
\dontrun{
stockSummaryTable_fun("Greater North Sea Ecoregion")
}
}
\author{
Scott Large
}
